Updates to adc calibration
Catch the case where no end of a calibration eye is found because the
adc clock rate is too low. Attempt to do something reasonable (guess where
the centre of the eye would be) based on clock speed.
Also a few changes to allow snap namees / verbosity options to be passed
to calibration methods.