Added script to fit core gains/delays/offsets
Basically this just calls methods from rww_tools. It allows the user
to input a sine wave of known frequency (specify this in the arguments
to the script accurately) and use it to calibrate the four adc cores.
Tested to the extent that spectra get better when it's used, but not
much beyone that. Values loosely agree with my own script that uses
FFTs to calculate calibration, though with my scripts (not yet checked in)
delays seem to be all over the place....