Hack to allow test mode to be turned on when zdok1 spi read interface fails.
For some compiles (not necessarily simply mapping to certain compile speeds or resource
utilisations) the SPI read interface on ZDOK1 seems to be faulty. In testing,
the ADC test mode could not be activated by reading the current config register,
modifying it, and writing the results back.
However, writing the appropriate value without a read seems to work. This commit
is a hideous, potentially dangerous workaround so that a control register modification
on ZDOK1 is performed by reading the control register from ZDOK0, modifying it's value, and
writing to ZDOK1.

I should absolutely work out why the read interface is failing.